added s3backer
s3backer implements a block level system on top of s3 this lets people use native FS's with a full set of all features instead of 'just' a filesystem emulation as fuse usually does and yieds a much better performance. The downside is that the number of files created and destroyed is astronomical and can bring down LeoFS due to having too many objects (the number of valid objects quickly reaches <1% w/o frequent compaction).